rpc: `getblocks` API (#141)
