+++ rpc/src/v1/traits/blockchain.rs
@@ -1,9 +1,7 @@
 use jsonrpc_core::Error;
 use jsonrpc_macros::Trailing;
 
-use v1::types::BlockchainInfo;
-use v1::types::GetBlockResponse;
-use v1::types::H256;
+use v1::types::{BlockMetadata, BlockchainInfo, GetBlockResponse, VerboseBlock, H256};
 
 build_rpc_trait! {
     /// Parity-randchain blockchain data interface.
@@ -38,5 +36,10 @@ build_rpc_trait! {
         /// @curl-example: curl --data-binary '{"jsonrpc": "2.0", "method": "getblockchaininfo", "id":1 }' -H 'content-type: application/json' http://127.0.0.1:8332/
         #[rpc(name = "getblockchaininfo")]
         fn blockchain_info(&self) -> Result<BlockchainInfo, Error>;
+
+        /// Get metadata of a number of blocks
+        /// @curl-example: curl --data-binary '{"jsonrpc": "2.0", "method": "getblocks", "params": [0, 10], "id":1 }' -H 'content-type: application/json' http://127.0.0.1:8332/
+        #[rpc(name = "getblocks")]
+        fn blocks(&self, u32, u32) -> Result<Vec<BlockMetadata>, Error>;
     }
 }
