+++ rpc/src/v1/types/mod.rs
@@ -3,19 +3,17 @@ mod block_template;
 mod block_template_request;
 mod blockchain;
 mod bytes;
-mod get_block_response;
 mod hash;
 mod network;
 mod nodes;
 mod submit_block;
 mod uint;
 
-pub use self::block::RawBlock;
+pub use self::block::{BlockMetadata, GetBlockResponse, RawBlock, VerboseBlock};
 pub use self::block_template::BlockTemplate;
 pub use self::block_template_request::{BlockTemplateRequest, BlockTemplateRequestMode};
 pub use self::blockchain::BlockchainInfo;
 pub use self::bytes::Bytes;
-pub use self::get_block_response::{GetBlockResponse, VerboseBlock};
 pub use self::hash::{H160, H256};
 pub use self::network::{Address, Network, NetworkInfo};
 pub use self::nodes::{AddNodeOperation, NodeInfo};
